requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
pandas==2.3.3
numpy==2.0.2
matplotlib==3.9.4
matplotlib-inline==0.2.1
scikit-learn==1.5.1
plotly==6.6.0
openai==2.29.0
python-dotenv==1.0.1
prophet==1.3.0
pydantic==2.12.5
SQLAlchemy==2.0.48
tqdm==4.67.3
psycopg2-binary
sqlalchemy